consultas/projeto_sisbelat/consulta_scr/requirements.txt
requests
pandas
openpyxl
beautifulsoup4
python-dotenv
